+++ custom_components/backup_checkup/sensor.py
@@ -221,6 +221,7 @@ SENSORS: tuple[BackupCheckupSensorDescription, ...] = (
             "content_inventory": data.recovery_content_inventory,
             "content_comparison": data.recovery_content_comparison,
             "storage_resilience": data.recovery_storage_resilience,
+            "preparedness": data.recovery_preparedness,
         },
     ),
     BackupCheckupSensorDescription(